+++ src/arena/mod.rs
@@ -7,7 +7,8 @@ pub mod directed;
 pub mod iterables;
 mod utils;
 
+pub use breadth::BreadthFirstIterator;
 pub use depth::{DepthFirstArenaTree, DepthFirstIterator};
-pub use directed::{ArenaNode, BreadthFirstIterator, DirectedArenaTree};
+pub use directed::{ArenaIndex, ArenaNode, DirectedArenaTree};
 use iterables::BaseDirectionIterable;
 pub use iterables::{BreadthFirstIterable, DepthFirstIterable, DirectionIterable, Nodelike};
